Move CliApplication to separate file
